+++ clients/core/src/Sidebar/InsideSidebar/components/InsideSidebarMenuItemCollapsable.tsx
